fix: move styles near the component
